Reverted previous fix and Reapplied in a correct way.
This new patch adds check's to individual RenderSingle calls to load the global script.  This is needed because using Html.Requires in MVC will  not call Render*Dependencies.

Also added a fix for web forms version having the same issue.  In this case the global scripts will be loaded so all that is need is to ensure that the script passed in is quoted.

As a added bonus I Updated the MVC LazyLoadRenderer to reader all its client side code in Render*Dependencies in one script tag instead of one for each. Ie:

<script>
CDLazyLoader.AddJs('//use.typekit.net/abc.js?cdv=10');
CDLazyLoader.AddJs('//site.two/file.js');
</script>